lex7_architecture/requirements.txt
# LEX-7 Architecture Dependencies
# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# State Space Models
mamba-ssm>=1.2.0
transformers>=4.35.0
accelerate>=0.24.0

# Distributed Computing & Communication
zenoh>=0.8.0
asyncio>=3.4.3
aioredis>=2.0.1

# Configuration & Data Management
pydantic>=2.4.0
pyyaml>=6.0.0
numpy>=1.24.0
pandas>=2.1.0

# Security & Cryptography
cryptography>=41.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0

# Testing & Validation
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Monitoring & Logging
tensorboard>=2.15.0
wandb>=0.15.0
structlog>=23.2.0

# Rust FFI
pyo3>=0.20.0
maturin>=1.4.0

# Development Tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
